zaia_microservice/pyproject.toml
[project]
name = "zaia_agents"
version = "0.1.0"
description = "zaia-agents using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai-tools>=0.36.0",
    "crewai[tools]>=0.102.0,<1.0.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "mem0ai>=0.1.60",
    "pymupdf>=1.25.3",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]

[project.scripts]
zaia_agents = "zaia_agents.main:run"
run_crew = "zaia_agents.main:run"
train = "zaia_agents.main:train"
replay = "zaia_agents.main:replay"
test = "zaia_agents.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
